+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 🧰 Bug fixes 🧰
 
+- Fix bug in ocb where flags did not take precedence. (#5726)
+
 ## v0.56.0 Beta
 
 ### 💡 Enhancements 💡