+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### 💡 Enhancements 💡
 
+- `ocb` now exits with an error if it fails to load the build configuration. (#5731)
+
 ### 🧰 Bug fixes 🧰
 
 ## v0.56.0 Beta
@@ -20,6 +22,7 @@
 - Expose `pcommon.NewSliceFromRaw` function (#5679)
 - `loggingexporter`: create the exporter's logger from the service's logger (#5677)
 - Add `otelcol_exporter_queue_capacity` metrics show the collector's exporter queue capacity (#5475)
+- Add support to handle 402, 413, 414, 431 http error code as permanent errors in OTLP exporter (#5685)
 
 ### 🧰 Bug fixes 🧰